まだ途中だけど色々改善
