+++ FlipByBlink/AppDelegate.swift
@@ -1,5 +1,5 @@
 import UIKit
-import PDFKit
+//import PDFKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -11,30 +11,30 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
-        print("🗣 url is ",url)
-        print("🗣 options is ",options)
-        
-        let ud = UIDocument(fileURL: url)
-        
-        print(ud)
-        
-        let 🗂 = FileManager.default
-        
-        let 📍 = URL(string: 🗂.urls(for: .documentDirectory, in: .userDomainMask)[0].absoluteString + "OpenedPDF.pdf")!
-
-        do{ try 🗂.removeItem(at: 📍)
-        }catch{ print("👿") }
-        
-        do{ try 🗂.copyItem(at: url, to: 📍)
-        }catch{ print("👿") }
-
-        if let vc:ViewController = window?.rootViewController as? ViewController{
-            if let 📘 = PDFDocument(url: 📍){
-                vc.📖.document = 📘
-                vc.📖.goToFirstPage(nil)
-                vc.📖.autoScales = true
-            }
-        }
+//        print("🗣 url is ",url)
+//        print("🗣 options is ",options)
+//
+//        let ud = UIDocument(fileURL: url)
+//
+//        print(ud)
+//
+//        let 🗂 = FileManager.default
+//
+//        let 📍 = URL(string: 🗂.urls(for: .documentDirectory, in: .userDomainMask)[0].absoluteString + "OpenedPDF.pdf")!
+//
+//        do{ try 🗂.removeItem(at: 📍)
+//        }catch{ print("👿") }
+//
+//        do{ try 🗂.copyItem(at: url, to: 📍)
+//        }catch{ print("👿") }
+//
+//        if let vc:ViewController = window?.rootViewController as? ViewController{
+//            if let 📘 = PDFDocument(url: 📍){
+//                vc.📖.document = 📘
+//                vc.📖.goToFirstPage(nil)
+//                vc.📖.autoScales = true
+//            }
+//        }
         return true
     }
 }
