+++ FlipByBlink/Assets.xcassets/bg.imageset/Contents.json
@@ -2,11 +2,11 @@
   "images" : [
     {
       "idiom" : "universal",
+      "filename" : "bg.pdf",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "bg.pdf",
       "scale" : "2x"
     },
     {
