+++ FlipByBlink/Assets.xcassets/fBBiCONrENEW.imageset/Contents.json
@@ -2,11 +2,11 @@
   "images" : [
     {
       "idiom" : "universal",
+      "filename" : "videoIcon.pdf",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "fBBiCONrENEWfOReMBED.pdf",
       "scale" : "2x"
     },
     {
