+++ FlipByBlink/Assets.xcassets/filerIconPdfForLaunch.imageset/Contents.json
@@ -2,11 +2,11 @@
   "images" : [
     {
       "idiom" : "universal",
+      "filename" : "filerIconPdfForLaunch.pdf",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "filerIconPdfForLaunch.pdf",
       "scale" : "2x"
     },
     {
