+++ FlipByBlink/Assets.xcassets/sampleIcon.imageset/Contents.json
@@ -2,11 +2,11 @@
   "images" : [
     {
       "idiom" : "universal",
+      "filename" : "sampleIcon.pdf",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "sampleIcon.pdf",
       "scale" : "2x"
     },
     {
